requirements.txt
fastapi==0.109.0
motor==3.3.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
uvicorn==0.27.0
pymongo==4.6.1
email-validator==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
